Update README.md
**Capture the Flag Competition (CTFs)**  
I am a professional bug hunter and I enjoy participating in Capture the Flag competitions. These days, I spend my time hunting for bugs in blockchain projects listed on [Immunefi](https://immunefi.com/). I have competed in WiCys and Google SANS Cyberstart competitions, Lockheed Martin Cyber Grand Challenge 2020, PNSQC 2020 Cyber Range, DTCC CTF (in collaboration with NYU CSAW'20), Hack-A-Sat 1 & 2 qualification round, and the National Cyber League 2020 Fall Season.

**Secure Coding**  
I had the pleasure of participating in the Hacker Halted Secure Coding Tournament and CTF in 2020 and 2021. The competition was a lot of fun, and it was great to see how well I matched up against other security professionals. I learned a lot during the tournament, and it was a great opportunity to improve my skills. I would recommend anyone who is interested in cybersecurity to take part in this competition.

**Conferences**  
I had the opportunity to attend both the 20th and 21st USENIX Security Conferences as well as the 17th Annual CSAW Cybersecurity Conference and Games. The presentations were extremely informative and provided insights into some of the latest cybersecurity threats and defenses. It was also great to see so many talented students participating in the CSAW Games competition. Overall, it was an excellent experience that I would highly recommend to anyone interested in computer security.